Corrected error on reading rgbd id from ud + tested 2 hands in mocap

+++ Bidirectional_interface/Assets/Scripts/UDPCommandManager.cs
@@ -67,7 +67,7 @@ public class UDPCommandManager : MonoBehaviour
 
             for (int i = 0; i < SimulationData.nbParametersMocap; i++)
             {
-                int id = System.BitConverter.ToInt32(data, (int)MocapIndices.id * 4);
+                int id = (int)System.BitConverter.ToSingle(data, (int)MocapIndices.id * 4);
                 
                 if (id == RightHandrigidbodyID)
                     optiTrackRigidbodies[(int)TrackedTargets.RightHand, i] = System.BitConverter.ToSingle(data, i * 4);
